525251 : add Java 9 module flag to classpath dependencies
Changes the Java configurator so that, if a Java project is a JPMS module, all classpath dependencies
which match required modules in module-info.java will have the module flage set to true.

One caveat is that, you currently need to call Maven Update project configuration after updating the
required modules in module-info.java, so the module flag can be added to the dependencies. See [2].

This change requires APIs from unreleased JDT (install from [1]) to compile, but m2e can still be installed
and run on older JDT versions.

[1] http://download.eclipse.org/eclipse/updates/4.7-Y-builds
[2] https://dev.eclipse.org/mhonarc/lists/m2e-dev/msg02068.html

Change-Id: I7e1a01bd5c57a262a7e9862d12a4a9c26ea203b1
Signed-off-by: Fred Bricon <fbricon@gmail.com>